+++ wip-domain/src/main/java/com/cvte/scm/wip/domain/core/requirement/util/ReqInsSplitHelper.java
@@ -112,7 +112,7 @@ public class ReqInsSplitHelper {
                     // 把剩余的数量都分配到最后一个行上
                     allocateQty = updateQty;
                 }
-                int resultQty = lotGroupLine.getReqQty() + allocateQty.intValue();
+                long resultQty = lotGroupLine.getReqQty() + allocateQty.longValue();
 
                 BigDecimal realAllocateQty = allocateQty;
                 if (resultQty < 0) {
@@ -138,7 +138,7 @@ public class ReqInsSplitHelper {
                 if (StringUtils.isBlank(lotGroupLine.getChangeType())) {
                     WipReqLineEntity reduceOrIncreaseLine = new WipReqLineEntity();
                     BeanUtils.copyProperties(lotGroupLine, reduceOrIncreaseLine);
-                    reduceOrIncreaseLine.setReqQty(realAllocateQty.abs().intValue())
+                    reduceOrIncreaseLine.setReqQty(realAllocateQty.abs().longValue())
                             .setUnitQty(realAllocateQty.abs().divide(wipLot.getLotQuantity(), 6, roundingMode).doubleValue());
                     if (InsOperationTypeEnum.REDUCE.getCode().equals(insDetail.getOperationType())) {
                         reduceOrIncreaseLine.setChangeType(ChangedTypeEnum.REDUCE.getCode());
